huge overhaul parts, starting equip/unequip overhaul (transfer functionality from item to backpack)

+++ java/com/felayga/unpixeldungeon/actors/buffs/FlavourBuff.java
@@ -37,7 +37,7 @@ public class FlavourBuff extends Buff {
 	//flavour buffs can all just rely on cooldown()
 	protected String dispTurns() {
 		//add one turn as buffs act last, we want them to end at 1 visually, even if they end at 0 internally.
-		float visualTurnsLeft = cooldown()+1f;
+		double visualTurnsLeft = cooldown()+1f;
 		return visualTurnsLeft == 1 ? "1 more turn" : new DecimalFormat("#.##").format(visualTurnsLeft) + " more turns";
 	}
 }
